requirements.txt
flask>=3.0.0
flask-restx>=1.3.0
flask-sqlalchemy>=3.1.1
flask-jwt-extended>=4.6.0
flask-bcrypt>=1.0.1
flask-limiter>=3.5.0
flask-migrate>=4.0.7
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=21.2.0
flask-cors>=4.0.0
